Add files via upload
In this project, we want to use data to get information about the relationship between social classes and social medias. To achieve the goal of the
project, we used two data sets, one of them is the dataset on www.kaggle.com
and the other one is collected data from a questionnaire.
After the investigations , we can say that wealthy strata spend more time
on social medias , and the weaker people spend less time on social medias.
Ofcourse , this difference isnot significant. in general , we can be said that
social medias are accessible for all social classes. The reasons of using social
medias in different classes are almost the same and the popular topics are not
much different in different classes.